backend/src/main/java/com/asdc/dalexchange/service/impl/CustomUserDetailsServiceImpl.java
package com.asdc.dalexchange.service.impl;

import com.asdc.dalexchange.model.User;
import com.asdc.dalexchange.repository.UserRepository;
import com.asdc.dalexchange.service.CustomUserDetailsService;
import lombok.AllArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

/**
 * Implementation of the CustomUserDetailsService interface for loading user-specific data.
 */
@Service
@Slf4j
@AllArgsConstructor
public class CustomUserDetailsServiceImpl implements CustomUserDetailsService {

    private UserRepository userRepository;

    /**
     * Loads the user by their email.
     *
     * @param email the email of the user to load
     * @return UserDetails containing the user's email, password, and roles
     * @throws UsernameNotFoundException if the user is not found
     */
    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        log.info("Attempting to load user by email: {}", email);

        // Retrieve the user by email from the repository
        User user = userRepository.findByEmail(email)
                .orElseThrow(() -> {
                    log.error("User not found with email: {}", email);
                    return new UsernameNotFoundException("User not found");
                });

        // Log the loaded user email (excluding the password for security reasons)
        log.debug("Loaded user with email: {}", user.getEmail());

        // Create and return a UserDetails object with the user's email, password, and roles
        return org.springframework.security.core.userdetails.User
                .withUsername(user.getEmail())
                .password(user.getPassword())
                .roles("USER")
                .build();
    }
}
